Move collision detection mixin to collisions directory
