+++ packages/flame/lib/game.dart
@@ -1,5 +1,6 @@
 /// {@canonicalFor text.TextPaint}
 /// {@canonicalFor text.TextRenderer}
+export 'src/collisions/has_collision_detection.dart';
 export 'src/extensions/vector2.dart';
 export 'src/game/camera/camera.dart';
 export 'src/game/camera/viewport.dart';
@@ -7,7 +8,6 @@ export 'src/game/flame_game.dart';
 export 'src/game/game_widget/game_widget.dart';
 export 'src/game/mixins/fps_counter.dart';
 export 'src/game/mixins/game.dart';
-export 'src/game/mixins/has_collision_detection.dart';
 export 'src/game/mixins/loadable.dart';
 export 'src/game/mixins/single_game_instance.dart';
 export 'src/game/projector.dart';
