+++ packages/flame/lib/src/collisions/has_collision_detection.dart
@@ -1,5 +1,5 @@
-import '../../../collisions.dart';
-import '../../../game.dart';
+import '../../collisions.dart';
+import '../../game.dart';
 
 /// Keeps track of all the [HitboxShape]s in the component tree and initiates
 /// collision detection every tick.
